--- oauthex/oauthex.go
@@ -1,14 +1,0 @@
-// Copyright 2025 The Go MCP SDK Authors. All rights reserved.
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file.
-
-// Package oauthex implements extensions to OAuth2.
-package oauthex
-
-import (
-	"github.com/modelcontextprotocol/go-sdk/internal/oauthex"
-)
-
-// ProtectedResourceMetadata is the metadata for an OAuth 2.0 protected resource,
-// as defined in section 2 of https://www.rfc-editor.org/rfc/rfc9728.html.
-type ProtectedResourceMetadata = oauthex.ProtectedResourceMetadata
